+++ src/main/java/nextstep/app/support/FormAuthentication.java
@@ -1,0 +1,29 @@
+package nextstep.app.support;
+
+import nextstep.app.ui.AuthenticationException;
+
+public class FormAuthentication implements Authentication {
+
+    private final boolean authenticated;
+    public FormAuthentication(final boolean authenticated) {
+        if (!authenticated) {
+            throw new AuthenticationException();
+        }
+        this.authenticated = true;
+    }
+
+    @Override
+    public Object getPrincipal() {
+        return null;
+    }
+
+    @Override
+    public Object getCredentials() {
+        return null;
+    }
+
+    @Override
+    public boolean isAuthenticated() {
+        return authenticated;
+    }
+}
